Split rendering to render.cc
